Renamed file for websocket protocol
