+++ mangum/protocols/__init__.py
@@ -1,10 +1,10 @@
 from .http import HTTPCycle
+from .websocket import WebSocketCycle
 from .lifespan import LifespanCycleState, LifespanCycle
-from .websockets import WebSocketCycle
 
 __all__ = [
     "HTTPCycle",
+    "WebSocketCycle",
     "LifespanCycleState",
     "LifespanCycle",
-    "WebSocketCycle",
 ]
